requirements.txt
jq
langchain_community
python-dotenv
langchain_pinecone
langchain_openai
langchain_core
langchain-google-genai
fastapi
mangum
uvicorn
psycopg2-binary